Update pitatebar
Rearranged the program. Dictionaries are now included in the functions that used them.
Removed global variables. The lists are now passed between the functions.
Amended the logic to keep asking the preferences until the user types "y" or "n".
Added the case in which the user answers no to all the questions.